requirements.txt
# Core agent orchestration
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# LLM & embeddings
openai>=1.50.0

# Vector store
chromadb>=0.5.0

# Web scraping & HTTP
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# PDF parsing (for research papers)
pymupdf>=1.24.0

# Data validation & schemas
pydantic>=2.8.0

# Report generation
python-docx>=1.1.0

# CLI interface
typer>=0.12.0
rich>=13.0.0

# Tracing & observability
langsmith>=0.1.0

# Web search (no API key needed)
duckduckgo-search>=6.3.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.5.0

# Web interface
streamlit>=1.38.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
